teste load data api na home
