+++ package.json
@@ -33,8 +33,10 @@
     "autoprefixer": "^10",
     "eslint": "^8",
     "eslint-config-next": "13.5.6",
+    "fetch-mock": "^9.11.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "jest-fetch-mock": "^3.0.3",
     "postcss": "^8",
     "tailwindcss": "^3",
     "ts-jest": "^29.1.1",
